advanced/submissions/team-members/art-turner/requirements.txt
# Core AI/LLM Dependencies
openai>=1.0.0
anthropic>=0.18.0

# Agent Frameworks
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# Vector Database & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Financial Data APIs
yfinance>=0.2.66
alpha-vantage>=2.3.1
finnhub-python>=2.4.19

# Web Search & Scraping
tavily-python>=0.3.0
google-search-results>=2.4.2
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# UI Framework
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
